feat: implemented useGetUsers hook
