+++ src/components/tutorial-to-me/UsersNotCompletedTutorialToMeTask.tsx
@@ -1,35 +1,8 @@
-'use client';
-import {useTeamStore} from '@/store/teamStore';
-import {useExercisesStore} from '@/store/useExercises';
-
-const getTeamMembersNotCompletedTask = (
-  teamMembers: {
-    id: string;
-    firstName: string | null;
-    lastName: string | null;
-  }[],
-  doneUsersId: string[],
-) => {
-  return teamMembers.filter((member) => !doneUsersId.includes(member.id));
-};
+import {useGetUsersById} from '@/hooks/useGetUsersById';
 
 const UsersNotCompletedTutorialToMeTask = () => {
-  const {currentTeam} = useTeamStore();
-  const {currentTutorialExerciseId, exercises} = useExercisesStore();
+  const {teamMembersNotCompletedTask} = useGetUsersById();
 
-  const teamMembers = currentTeam?.team_members.map((member) => ({
-    id: member.user_id,
-    firstName: member.first_name,
-    lastName: member.last_name,
-  }));
-
-  const doneUsersId = exercises
-    .filter((e) => e.exerciseId === currentTutorialExerciseId)
-    .map((e) => e.replied_id);
-
-  const teamMembersNotCompletedTask = teamMembers
-    ? getTeamMembersNotCompletedTask(teamMembers, doneUsersId)
-    : [];
   return (
     <div className="flex">
       {teamMembersNotCompletedTask.length > 0
